Import Format API components instead of passing as props
