+++ packages/editor/src/components/rich-text/format-edit.js
@@ -1,91 +1,14 @@
 /**
  * WordPress dependencies
  */
-import { normalizeIconObject } from '@wordpress/blocks';
 import { withSelect } from '@wordpress/data';
-import { Component, Fragment } from '@wordpress/element';
+import { Fragment } from '@wordpress/element';
 import { getActiveFormat } from '@wordpress/rich-text';
-import { Fill, KeyboardShortcuts, ToolbarButton } from '@wordpress/components';
-import { rawShortcut, displayShortcut } from '@wordpress/keycodes';
-
-/**
- * Internal dependencies
- */
-import InserterListItem from '../inserter-list-item';
-import { normalizeTerm } from '../inserter/menu';
-
-function isResult( { title, keywords = [] }, filterValue ) {
-	const normalizedSearchTerm = normalizeTerm( filterValue );
-	const matchSearch = ( string ) => normalizeTerm( string ).indexOf( normalizedSearchTerm ) !== -1;
-	return matchSearch( title ) || keywords.some( matchSearch );
-}
-
-function FillToolbarButton( { name, shortcutType, shortcutCharacter, ...props } ) {
-	let shortcut;
-	let fillName = 'RichText.ToolbarControls';
-
-	if ( name ) {
-		fillName += `.${ name }`;
-	}
-
-	if ( shortcutType && shortcutCharacter ) {
-		shortcut = displayShortcut[ shortcutType ]( shortcutCharacter );
-	}
-
-	return (
-		<Fill name={ fillName }>
-			<ToolbarButton
-				{ ...props }
-				shortcut={ shortcut }
-			/>
-		</Fill>
-	);
-}
-
-function FillInserterListItem( props ) {
-	return (
-		<Fill name="Inserter.InlineElements">
-			{ ( { filterValue } ) => {
-				if ( filterValue && ! isResult( props, filterValue ) ) {
-					return null;
-				}
-
-				return <InserterListItem { ...props } icon={ normalizeIconObject( props.icon ) } />;
-			} }
-		</Fill>
-	);
-}
-
-class Shortcut extends Component {
-	constructor() {
-		super( ...arguments );
-
-		this.onUse = this.onUse.bind( this );
-	}
-
-	onUse() {
-		this.props.onUse();
-		return false;
-	}
-
-	render() {
-		const { character, type } = this.props;
-
-		return (
-			<KeyboardShortcuts
-				bindGlobal
-				shortcuts={ {
-					[ rawShortcut[ type ]( character ) ]: this.onUse,
-				} }
-			/>
-		);
-	}
-}
 
 const FormatEdit = ( { formatTypes, onChange, value } ) => {
 	return (
 		<Fragment>
-			{ formatTypes.map( ( { name, edit: Edit, keywords } ) => {
+			{ formatTypes.map( ( { name, edit: Edit } ) => {
 				if ( ! Edit ) {
 					return null;
 				}
@@ -101,14 +24,6 @@ const FormatEdit = ( { formatTypes, onChange, value } ) => {
 						activeAttributes={ activeAttributes }
 						value={ value }
 						onChange={ onChange }
-						ToolbarButton={ FillToolbarButton }
-						InserterListItem={ ( props ) =>
-							<FillInserterListItem
-								keywords={ keywords }
-								{ ...props }
-							/>
-						}
-						Shortcut={ Shortcut }
 					/>
 				);
 			} ) }
