+++ packages/format-library/src/bold/index.js
@@ -4,6 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
 import { toggleFormat } from '@wordpress/rich-text';
+import { RichTextToolbarButton, RichTextShortcut } from '@wordpress/editor';
 
 const name = 'core/bold';
 
@@ -13,17 +14,17 @@ export const bold = {
 	match: {
 		tagName: 'strong',
 	},
-	edit( { isActive, value, onChange, ToolbarButton, Shortcut } ) {
+	edit( { isActive, value, onChange } ) {
 		const onToggle = () => onChange( toggleFormat( value, { type: name } ) );
 
 		return (
 			<Fragment>
-				<Shortcut
+				<RichTextShortcut
 					type="primary"
 					character="b"
 					onUse={ onToggle }
 				/>
-				<ToolbarButton
+				<RichTextToolbarButton
 					name="bold"
 					icon="editor-bold"
 					title={ __( 'Bold' ) }
