+++ packages/format-library/src/code/index.js
@@ -2,8 +2,8 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { Fragment } from '@wordpress/element';
 import { toggleFormat } from '@wordpress/rich-text';
+import { RichTextShortcut } from '@wordpress/editor';
 
 const name = 'core/code';
 
@@ -13,17 +13,15 @@ export const code = {
 	match: {
 		tagName: 'code',
 	},
-	edit( { value, onChange, Shortcut } ) {
+	edit( { value, onChange } ) {
 		const onToggle = () => onChange( toggleFormat( value, { type: name } ) );
 
 		return (
-			<Fragment>
-				<Shortcut
-					type="access"
-					character="x"
-					onUse={ onToggle }
-				/>
-			</Fragment>
+			<RichTextShortcut
+				type="access"
+				character="x"
+				onUse={ onToggle }
+			/>
 		);
 	},
 };
