+++ packages/format-library/src/image/index.js
@@ -5,7 +5,7 @@ import { Path, SVG } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { Fragment, Component } from '@wordpress/element';
 import { insertObject } from '@wordpress/rich-text';
-import { MediaUpload } from '@wordpress/editor';
+import { MediaUpload, RichTextInserterListItem } from '@wordpress/editor';
 
 const ALLOWED_MEDIA_TYPES = [ 'image' ];
 
@@ -44,11 +44,12 @@ export const image = {
 		}
 
 		render() {
-			const { value, onChange, InserterListItem } = this.props;
+			const { value, onChange } = this.props;
 
 			return (
 				<Fragment>
-					<InserterListItem
+					<RichTextInserterListItem
+						name={ name }
 						icon={ <SVG xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><Path d="M4 16h10c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2H4c-1.1 0-2 .9-2 2v9c0 1.1.9 2 2 2zM4 5h10v9H4V5zm14 9v2h4v-2h-4zM2 20h20v-2H2v2zm6.4-8.8L7 9.4 5 12h8l-2.6-3.4-2 2.6z" /></SVG> }
 						title={ __( 'Inline Image' ) }
 						onClick={ this.openModal }
