+++ packages/format-library/src/italic/index.js
@@ -4,6 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
 import { toggleFormat } from '@wordpress/rich-text';
+import { RichTextToolbarButton, RichTextShortcut } from '@wordpress/editor';
 
 const name = 'core/italic';
 
@@ -13,17 +14,17 @@ export const italic = {
 	match: {
 		tagName: 'em',
 	},
-	edit( { isActive, value, onChange, ToolbarButton, Shortcut } ) {
+	edit( { isActive, value, onChange } ) {
 		const onToggle = () => onChange( toggleFormat( value, { type: name } ) );
 
 		return (
 			<Fragment>
-				<Shortcut
+				<RichTextShortcut
 					type="primary"
 					character="i"
 					onUse={ onToggle }
 				/>
-				<ToolbarButton
+				<RichTextToolbarButton
 					name="italic"
 					icon="editor-italic"
 					title={ __( 'Italic' ) }
