+++ packages/format-library/src/strikethrough/index.js
@@ -4,6 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
 import { toggleFormat } from '@wordpress/rich-text';
+import { RichTextToolbarButton, RichTextShortcut } from '@wordpress/editor';
 
 const name = 'core/strikethrough';
 
@@ -13,17 +14,17 @@ export const strikethrough = {
 	match: {
 		tagName: 'del',
 	},
-	edit( { isActive, value, onChange, ToolbarButton, Shortcut } ) {
+	edit( { isActive, value, onChange } ) {
 		const onToggle = () => onChange( toggleFormat( value, { type: name } ) );
 
 		return (
 			<Fragment>
-				<Shortcut
+				<RichTextShortcut
 					type="access"
 					character="d"
 					onUse={ onToggle }
 				/>
-				<ToolbarButton
+				<RichTextToolbarButton
 					name="strikethrough"
 					icon="editor-strikethrough"
 					title={ __( 'Strikethrough' ) }
